+++ CHANGELOG.md
@@ -11,6 +11,7 @@ develop
     - Now, “Find All” action also highlights all matched strings in the editor, and thereby “Highlight” action is removed.
     - On Yosemite and later, a visual feedback is shown when the seach wrapped.
     - Keep selected range after “Replace All” with in-selection option.
+    - Display total number of found in find panel on simple find actions.
     - Now, “Find All” and “Replace All” actions are able to process multiple selections.
     - Revert “Use selection for Replace” action to allow using an empty string.
     - Tweak layout and style.